#35 [mod] Modify mapview by mapfragment
